refactor(team): extract log source watcher ignore and pending sessions
Two self-contained concerns move out of TeamLogSourceTracker.ts into
sibling modules; no behaviour changes.

teamLogSourceWatcherIgnore.ts takes shouldIgnoreLogSourceWatcherPath
verbatim. The function body is byte-identical to its old form; only the
imports it needs from teamLogSourceWatchScope travel with it. Two of
those specifiers (BOARD_TASK_CHANGES_DIRNAME, getRelativeLogSourceParts)
had no other reader in the tracker and are dropped from its import list.

teamLogSourcePendingSessions.ts takes the five pending-unknown-session
helpers that were private methods on TeamLogSourceTracker. None of them
touched `this` except to call each other, so they become module
functions over an explicit PendingUnknownSessionState port describing
the two TrackingState fields they read; TrackingState satisfies it
structurally, so the five call sites change only by losing the `this.`
prefix. prunePendingUnknownSessions stays module-private -- it was
private before and nothing outside the group calls it. The state port
is not exported either, so the module's public surface is exactly the
four helpers plus the PendingUnknownSessionCandidate shape the tracker
still needs for TrackingState. MAX_PENDING_UNKNOWN_ROOT_REFRESH_ATTEMPTS,
MAX_PENDING_UNKNOWN_ROOT_SESSIONS and PENDING_UNKNOWN_ROOT_SESSION_TTL_MS
move with them and leave the tracker's import list.

The one existing test that drove shouldIgnoreLogSourceWatcherPath
directly used no shared fixture, so it moves verbatim into a colocated
teamLogSourceWatcherIgnore.test.ts; its 14 assertions are unchanged. The
tracker's own suite keeps its other 13 cases untouched apart from
collapsing the now single-symbol import.

The near-duplicate ignore logic inside classifyLogSourceWatcherEvent in
teamLogSourceWatchScope.ts is deliberately left alone -- folding the two
together would be a behaviour change, not a move.

TeamLogSourceTracker.ts: 1088 -> 970 lines, against a frozen cap of 1088.
The baseline JSON is unchanged.
